loading menu items dynamically from the database based on the route location pathname

+++ frontend/src/components/Layout/Layout.jsx
@@ -1,23 +1,43 @@
-import './Layout.css'
-import Menu from '../Menu/Menu'
-import { Outlet, useLocation } from 'react-router'
-import { useEffect } from 'react'
+import "./Layout.css";
+import Menu from "../Menu/Menu";
+import { Outlet, useLocation, useNavigate } from "react-router";
+import { useEffect, useState } from "react";
+import axios from "axios";
 
-function Layout(){
+function Layout() {
+  const location = useLocation();
+  const navigate = useNavigate();
 
-    const location = useLocation()
+  const [menuItems, setMenuItems] = useState([]);
 
-    useEffect(() => {
-        console.log("clearing local storage...")
-        window.localStorage.clear()
-    }, [location.pathname == '/login' || location.pathname == '/logout'])
+  useEffect(() => {
+    console.log("running use effect")
+    if ((location.pathname == "/login", location.pathname == "/logout")) {
+      console.log("clearing local storage...");
+      console.log(location.pathname);
+      window.localStorage.clear();
+    }
 
-    return <div className='layout-container'>
-        <Menu />
-        <div className='content-container'>
-            <Outlet />
-        </div>
+    axios.get("http://localhost:8080/api/v1/menu-items", window.localStorage.getItem("token") ? {headers: {"Authorization": `Bearer ${window.localStorage.getItem("token")}`}} : null)
+        .then(res => {
+            console.log(res.data)
+            setMenuItems(res.data)
+        })
+        .catch(err => {
+            if(err.status == 401 || err.status == 403){
+                navigate('/login')
+            }
+        })
+  }, [location]);
+
+  return (
+    <div className="layout-container">
+      <Menu menuItems={menuItems}/>
+      <div className="content-container">
+        <Outlet />
+      </div>
     </div>
+  );
 }
 
-export default Layout
+export default Layout;
